pyproject.toml
[project]
name = "0-hitl"
version = "0.1.0"
description = "Zero Human In The Loop - Agent OS"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "pydantic>=2.0.0",
    "litellm",
    "tiktoken",
    "docker",
    "aiosqlite",       
    "python-dotenv",
    "pyyaml",
    "httpx"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["core", "gateway"]

[tool.hatch.build.targets.sdist]
include = [
    "/core",
    "/gateway",
    "/profiles",
    "/skills",
    "/main.py",
    "/README.md",
    "/pyproject.toml",
]
